tty/serial/8250 - 16550a FIFO Control sysfs
    Implements the fifo_control uart_ops into 8250 driver for 16550a compatible devices.
A dispatch function retrieves the  fifo_control callbacks stored in the uart_config[] for
the specified port type. Most port types can opt into the default 16550a callback to offer
minimal fifo control via the FCR register (similar to the prior rx_trig_bytes sysfs),
OR developers may opt to implement their own callbacks for more fine-grained fifo control.

    * Add 8250_fifo:
         - Implement dispatch_set_fifo_control
           dispatched function based on port type

    * Define fifo control callbacks in uart_8250_ops:
         - serial8250_get_fifo_control
         - serial8250_set_fifo_control

      NOTABLE CHANGES:
          - fifo_control is intended to become the canonical representation of UART fifo state,
eventually deprecating the role of FCR and other registers in 8250_port. The logic is that fifo state
isn't necesarrily nor completely captured by FCR register on later model serial devices. For the time being,
we MUST ensure the states in FCR and the fifo_control are the same.
Any changes to FCR MUST be reflected in fifo_control.

          -  Although the sysfs entry for rx_trig_bytes is backwards compatible by name, the behaviour of the sysfs IS NOT
backwards compatible. The new rx_trig_bytes does NOT automatically decrement to the next lowest value available. If the specified value is not
available, the driver will NOT change the fifo trigger.