tty/serial/8250 - 16550a FIFO Control sysfs
Implements the fifo_control uart_ops into 8250_port.c for 16550a compatible devices.

    - Define fifo control callbacks in uart_8250_ops:
         * serial8250_get_fifo_control
         * serial8250_set_fifo_control

    - Implement serial8250_dispatch_set_fifo_control to support future FIFO
control modes (eg. 16650, etc).

 NOTABLE CHANGES:
    - fifo_control is intended to become the canonical representation of UART fifo state,
eventually deprecating the role of FCR in 8250_port. The logic is that fifo state
isn't necesarrily nor completely captured by FCR register on later model serial devices. For the time being,
we MUST ensure the states in FCR and the fifo_control are the same.
Any changes to FCR MUST be reflected in fifo_control.

    - The current rx_trig_bytes sysfs implementation deviates from older
implementations in that it does not automatically select the next lowest level.